scripts: act: better debug
We were actually running wget and letting the user see the output...
and get its error code, which is quite fine in a manual run, but
messes with the automatic running of these scripts.

Now, we still do wget and validate its error code, but instead the
script then does its business as usual (and hopefully this fixes our
cron'ed GA).

This patch also changes the time at which the scheduled cronjob runs,
just because it is still once a day, which is what we want, and will
give us still a run today (so we can have it run with this change
sooner rather than later).